pyproject.toml
[project]
name = "second-brain-chat"
version = "0.1.0"
description = "LLM-powered chat interface for navigating second-brain Freeplane mindmaps"
authors = [{name = "Joey Corea", email = "joeycorea@gmail.com"}]
readme = "README.md"
packages = [{ include = "second_brain_chat" }]
requires-python = "^3.10"

# Core runtime dependencies
dependencies = [
    "langchain>=0.1.16,<0.2.0",
    "langchain-openai>=0.0.8,<0.1.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "chromadb>=0.4.15,<0.5.0",
    "openai>=1.14.2",
    "torch>=2.0.0,<3.0.0",
    "sentence-transformers>=2.2.2,<3.0.0",
    "tiktoken>=0.5.1,<0.6.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0"
]

# Developer dependencies (install with --dev flag)
[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
ruff = "^0.3.4"
mypy = "^1.9.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

# Poetry build backend
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
